% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{check_rlernen_update}
\alias{check_rlernen_update}
\title{Check if a new commit of rlernen is available on GitHub}
\usage{
check_rlernen_update()
}
\value{
message whether update is available
}
\description{
Check if a new commit of rlernen is available on GitHub
}
