% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grading.R
\name{grade_auto}
\alias{grade_auto}
\title{rlernen's autograder}
\usage{
grade_auto(
  success_message = "Richtig!",
  generic_fail_message = "Noch nicht ganz richtig.",
  tol = 1e-08,
  env = parent.frame()
)
}
\arguments{
\item{success_message}{Message to display when the user's solution is correct.}

\item{generic_fail_message}{Message to display when the user's solution is incorrect, without specific hints.}

\item{tol}{Numeric tolerance for comparing numeric results (default: 1e-8).}

\item{env}{The environment containing the necessary objects for grading (default: parent.frame())}
}
\description{
Grades the user's solution with automatic hint generation based on the solution code.
}
