% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{homework}
\alias{homework}
\title{Open homework assignment}
\usage{
homework(number, overwrite = F)
}
\arguments{
\item{number}{numeric value of homework assignment}

\item{overwrite}{boolean should existing homework file in current folder be
overwritten?}
}
\value{
side effect: opens homework file in Rstudio or R
}
\description{
Open homework assignment
}
