% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grading.R
\name{infer_suggestion_map_from_solution}
\alias{infer_suggestion_map_from_solution}
\title{Infer a suggestion map from the solution code.}
\usage{
infer_suggestion_map_from_solution(
  solution_code,
  keep_values = TRUE,
  eval_env = baseenv()
)
}
\arguments{
\item{solution_code}{The solution code as a character string.}

\item{keep_values}{Whether to keep the evaluated argument values in the suggestion map (default)}

\item{eval_env}{The environment in which to evaluate argument expressions for value extraction (default: baseenv())}
}
\description{
This function parses the solution code to identify which functions and arguments are used in the solution. It constructs a suggestion map that can be used to provide targeted hints to the user based on their code.
}
