% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{knit_print.hsearch}
\alias{knit_print.hsearch}
\title{Methods for printing help search results, package info, and data/vignette}
\usage{
\method{knit_print}{hsearch}(x, ...)
}
\arguments{
\item{x}{An R object to be printed}

\item{...}{Additional arguments passed to knit_print()
help.search()}
}
\description{
Methods for printing help search results, package info, and data/vignette
}
