% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{knit_print.packageInfo}
\alias{knit_print.packageInfo}
\title{Methods for printing help search results, package info, and data/vignette}
\usage{
\method{knit_print}{packageInfo}(x, ...)
}
\arguments{
\item{x}{An R object to be printed}

\item{...}{Additional arguments passed to knit_print()
library(help = foo)}
}
\description{
Methods for printing help search results, package info, and data/vignette
}
