% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{knit_print.packageIQR}
\alias{knit_print.packageIQR}
\title{Methods for printing help search results, package info, and data/vignette}
\usage{
\method{knit_print}{packageIQR}(x, options, ...)
}
\arguments{
\item{x}{An R object to be printed}

\item{options}{A list of options passed to knit_print()}

\item{...}{Additional arguments passed to knit_print()
data()/vignette()}
}
\description{
Methods for printing help search results, package info, and data/vignette
}
