% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{q_n}
\alias{q_n}
\title{Create a numeric question with retry allowed}
\usage{
q_n(answer, text = "Was kommt heraus?")
}
\arguments{
\item{answer}{numeric answer to the question}

\item{text}{question text}
}
\description{
Create a numeric question with retry allowed
}
