% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{question}
\alias{question}
\title{Create a question with retry allowed}
\usage{
question(...)
}
\arguments{
\item{...}{question components (e.g., answer(...))}
}
\description{
Create a question with retry allowed
}
