% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlernen-package.R
\docType{package}
\name{rlernen-package}
\alias{rlernen}
\alias{rlernen-package}
\title{rlernen: learnr Tutorial for Research Methods (German)}
\description{
German learnr tutorials for research methods, covering topics such as data wrangling, descriptive statistics, basic plots, statistical tests, regression, factor analysis and contrast analysis. A hosted version is available at \url{https://rlernen.de}.
}
\author{
\strong{Maintainer}: Johannes Titz \email{johannes@titz.science} (\href{https://orcid.org/0000-0002-1102-5719}{ORCID})

}
\keyword{internal}
