% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{homework}
\alias{homework}
\title{Open homework assignment}
\usage{
homework(number)
}
\arguments{
\item{number}{numeric value of homework assignment}
}
\value{
side effect: opens homework file in Rstudio or R
}
\description{
Open homework assignment
}
